Integration tests: Add a test of encoding an empty vector
This test fails for v0.2.3. It encodes to [128], which is definitely
wrong. That value of 128 seems to come from the `test_pair()` function
initialising the output array to 0x80 before the test. So I guess the
`encode()` function is setting the output length to 1 but not writing
any value.

It's arguable whether it should encode to [1] or []. I prefer [1], and I
implemented that for my Python and C COBS code. In a comms protocol it
allows "empty frame" to be distinguished from "nothing at all", which
can be useful for some protocols.